+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Yorc Changelog
 
-## UNRELEASED
+## 4.0.3 (August 27, 2020)
 
 ### FEATURES